Resolving issue with error params encapsulated as Promise in next
