+++ app/page.tsx
@@ -1,0 +1,9 @@
+import MainDashboard from "./dashboard/page";
+
+export default function Home() {
+  return (
+    <main className="p-10 fixed inset-0 flex flex-col items-center justify-center">
+      <MainDashboard />
+    </main>
+  );
+}
