+++ prisma/schema.prisma
@@ -1,0 +1,48 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+// Looking for ways to speed up your queries, or scale easily with your serverless or edge functions?
+// Try Prisma Accelerate: https://pris.ly/cli/accelerate-init
+
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABASE_URL")
+  directUrl = env("DIRECT_URL")
+}
+
+model Tenant {
+  id String @id @default(cuid())
+  name String
+  slug String @unique
+  memberships Membership[]
+  createdAt DateTime @default(now())
+}
+
+model User {
+  id String @id @default(cuid())
+  clerkUserId String @unique
+  email String
+  memberships Membership[]
+  createdAt DateTime @default(now())
+}
+
+model Membership {
+  id String @id @default(cuid())
+  userId String
+  tenantId String
+  role Role @default(MEMBER)
+  user User @relation(fields: [userId], references: [id])
+  tenant Tenant @relation(fields: [tenantId], references: [id])
+
+  @@unique([userId,tenantId])
+}
+
+enum Role {
+  OWNER
+  ADMIN
+  MEMBER
+}
