--- src/app/api/onboard/route.ts
@@ -1,8 +1,0 @@
-import { onboardUser } from "@/lib/onboardUser";
-import { NextResponse } from "next/server";
-
-export async function POST(request: Request) {
-  const body = await request.json();
-  const { tenantSlug } = await onboardUser(body);
-  return NextResponse.json({ tenantSlug }, { status: 200 });
-}
