--- src/app/layout.tsx
@@ -1,27 +1,0 @@
-import { ClerkProvider } from "@clerk/nextjs";
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "./globals.css";
-
-const inter = Inter({
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "SAAS Boilerplate",
-  description: "SAAS boilerplate with multi-tenancy",
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <ClerkProvider>
-      <html lang="en">
-        <body className={`${inter.className} antialiased`}>{children}</body>
-      </html>
-    </ClerkProvider>
-  );
-}
