--- src/app/page.tsx
@@ -1,9 +1,0 @@
-import Dashboard from "./dashboard/page";
-
-export default function Home() {
-  return (
-    <main className="p-10 fixed inset-0 flex flex-col items-center justify-center">
-      <Dashboard />
-    </main>
-  );
-}
